Added 3 command and made some performance changes. The added commands are exec, eval and update command and removed the bot variable from most of the files to lower ram usage. Added more and correct types to files.

+++ src/handlers/CommandHandler.ts
@@ -102,12 +102,16 @@ export default class CommandHandler {
                     //     };
                     // }
 
-                    if (command.meta.deleteMessage === true)
-                        await message.delete();
-
                     const commandArgs = [...args];
                     commandArgs.shift();
                     const ctx = new Context(this.bot, message, commandArgs);
+
+                    if (
+                        command.meta.deleteMessage === true &&
+                        ctx.me.permission.has("manageMessages")
+                    )
+                        await message.delete();
+
                     await command.execute(ctx, this.base);
 
                     return true;
